fix: dead code elimination
This fixes an issue where it naively tried to remove imports
by assuming they could all go.

It's not bulletproof because I ran into a weird issue.
Babel 7 is not updating the entire tree so I am
tracking the removed calls. That doesn't feel right but it
passes all of the test cases.